+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 [19.02.2023]:
+=======================================================================================================================================================================
 - New web app design;
 - Added "blank search input" block;
 - Added "not found" block;
@@ -8,6 +9,7 @@
 wind speed;
 
 [21.02.2023]:
+=======================================================================================================================================================================
 - Added "clear search input" button;
 - Added new information blocks: visibility, cloudiness, air quality
 and pressure;
@@ -20,11 +22,13 @@ temperature types, now non functional);
 - Removed country flag icons from "weather main" block;
 
 [22.02.2023]:
+=======================================================================================================================================================================
 - Fixed "description" info displaying;
 - Added functionality to "clear search input" button;
 - Added date and time blocks to web app;
 
 [23.02.2023]:
+=======================================================================================================================================================================
 - Added country flag icon to "weather main" block;
 - Added "save location" button;
 - Added scaling animations to all buttons;
@@ -34,12 +38,14 @@ temperature types, now non functional);
 - Added "delete saved location" button;
 
 [24.02.2023]:
+=======================================================================================================================================================================
 - Removed all flags images from project;
 - Integrated "FlagsAPI" api to get any country flag from 2-alpha country code;
 - A check for an identical location in the list "Saved locations" was added 
 to exclude the possibility of duplication of these values;
 
 [25.02.2023]:
+=======================================================================================================================================================================
 - Fixed issue with saving location (The location was saved with 
 the value from the input and not with the value of the location 
 that the user is looking for);
@@ -48,11 +54,13 @@ that the user is looking for);
 - Added hover animations to "highlight" item;
 
 [26.02.2023]:
+=======================================================================================================================================================================
 - Integrated "String.js" library (Strings);
 - Added functional to "Celsium" and "Fahrenheit" buttons:
     Now users can switch temperature type between celsium and fahrenheit types;
 
 [27.02.2023]:
+=======================================================================================================================================================================
 - Added "Copy" buttons to all "weather highlight" blocks:
     Now users can copy any weather detail to clipboard and share this information to each other;
 - Added saving system (LocalStorage):
@@ -60,5 +68,13 @@ that the user is looking for);
     all saved locations;
 
 [01.03.2023]:
+=======================================================================================================================================================================
 - Integrated "JQuery" library;
 - Fixed issues with saving and removing location;
+
+[04.03.2023]:
+=======================================================================================================================================================================
+- Redesigned "header" block;
+- Redesigned all "header" buttons;
+- Redesigned all "highlight" items;
+- Added "search history" button;